mex matrix updated

+++ src/tutorial/helmholtz_bem_2d_cyl.mex.cpp
@@ -43,8 +43,8 @@ void mexFunction(int nlhs, mxArray *lhs[], int nrhs, mxArray const *rhs[])
 //! [Function spaces]
 
 //! [Matrices]
-	int n = surf_sp.get_num_dofs();
-	int m = field_sp.get_num_dofs();
+	size_t n = surf_sp.get_num_dofs();
+	size_t m = field_sp.get_num_dofs();
 	cMatrix Ls(n, n, lhs[0]), Lf(m, n, lhs[2]);
 	cMatrix Ms(n, n, lhs[1]), Mf(m, n, lhs[3]);
 //! [Matrices]
