+++ src/tutorial/helmholtz_bem_3d_bm.mex.cpp
@@ -16,7 +16,7 @@ void mexFunction(int nlhs, mxArray *lhs[], int nrhs, mxArray const *rhs[])
 	
 	auto const &surf_sp = NiHu::constant_view(surf_mesh);
 
-	int n = surf_sp.get_num_dofs();
+	size_t n = surf_sp.get_num_dofs();
 	cMatrix L_surf(n, n, lhs[0]), M_surf(n, n, lhs[1]),
 		Mt_surf(n, n, lhs[2]), N_surf(n, n, lhs[3]);
