+++ src/tutorial/helmholtz_bem_3d_full.mex.cpp
@@ -16,8 +16,8 @@ void mexFunction(int nlhs, mxArray *lhs[], int nrhs, mxArray const *rhs[])
 	auto const &surf_sp = NiHu::constant_view(surf_mesh);
 	auto const &field_sp = NiHu::constant_view(field_mesh);
 
-	int n = surf_sp.get_num_dofs();
-	int m = field_sp.get_num_dofs();
+	size_t n = surf_sp.get_num_dofs();
+	size_t m = field_sp.get_num_dofs();
 	cMatrix Ls(n, n, lhs[0]), Ms(n, n, lhs[1]), Mts(n, n, lhs[2]), Ds(n, n, lhs[3]),
 		Gxxs(n, n, lhs[4]), Lf(m, n, lhs[5]), Mf(m, n, lhs[6]), Mtf(m, n, lhs[7]);
