+++ src/tutorial/laplace_bem_2d_full.mex.cpp
@@ -13,11 +13,11 @@ void mexFunction(int nlhs, mxArray *lhs[], int nrhs, mxArray const *rhs[])
 	auto field_mesh = NiHu::create_mesh(field_nodes, field_elem, NiHu::line_1_tag());
 	auto const &surf_sp = NiHu::constant_view(surf_mesh);
 	auto const &field_sp = NiHu::constant_view(field_mesh);
-	int n = surf_sp.get_num_dofs();
+	size_t n = surf_sp.get_num_dofs();
 	dMatrix L_surf(n, n, lhs[0]);
 	dMatrix M_surf(n, n, lhs[1]);
 	dMatrix W_surf(n, n, lhs[2]);
-	int m = field_sp.get_num_dofs();
+	size_t m = field_sp.get_num_dofs();
 	dMatrix L_field(m, n, lhs[3]);
 	dMatrix M_field(m, n, lhs[4]);
 	auto L = NiHu::create_integral_operator(NiHu::laplace_2d_SLP_kernel());
